381 refactor headerlogotsx (#392)
* refactor HeaderLogo.tsx

* comments data
