+++ src/assets/styles/variables.scss
@@ -72,8 +72,8 @@
   --footerChip-closeIcon-size: 21px;
   --fullscreenButton-size: 30px;
   --header-container-height: calc(var(--header-height) * 1px);
-  --headerLogo-height: var(--header-container-height);
-  --headerLogo-width: 130px;
+  --helpInfo-height: var(--header-container-height);
+  --helpInfo-width: 130px;
   --pagination-button-size: 38px;
   --pagination-buttonWrap-height: 48px;
   --pagination-buttonWrap-width: 38px;
