+++ src/components/Credits/Credits.module.scss
@@ -1,0 +1,6 @@
+.credits {
+  font-size: var(--text-3xl);
+  padding-left: var(--padding-7);
+  padding-right: var(--padding-7);
+  z-index: 5000;
+}
