--- src/components/HeaderLogo/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HeaderLogo';
