+++ src/components/HelpInfo/HelpInfo.module.scss
@@ -1,12 +1,12 @@
-.headerLogo {
+.helpInfo {
   position: absolute;
   top: -1px; // cover up top app border
   right: 0;
   cursor: pointer;
   border: none;
-  height: var(--headerLogo-height);
+  height: var(--helpInfo-height);
   text-align: center;
-  width: var(--headerLogo-width);
+  width: var(--helpInfo-width);
   font-size: var(--text-2xl);
   background: var(--red-200);
   color: var(--white);
